fixed KL computation and eased logging

+++ examples/sac/main.py
@@ -29,11 +29,11 @@ parser.add_argument('--automatic_entropy_tuning', type=bool, default=True, metav
                     help='Automaically adjust α (default: False)')
 parser.add_argument('--seed', type=int, default=123456, metavar='N',
                     help='random seed (default: 123456)')
-parser.add_argument('--batch_size', type=int, default=256, metavar='N',
+parser.add_argument('--batch_size', type=int, default=128, metavar='N',
                     help='batch size (default: 256)')
 parser.add_argument('--num_steps', type=int, default=2000001, metavar='N',
                     help='maximum number of steps (default: 1000000)')
-parser.add_argument('--hidden_size', type=int, default=128, metavar='N',
+parser.add_argument('--hidden_size', type=int, default=256, metavar='N',
                     help='hidden size (default: 256)')
 parser.add_argument('--updates_per_step', type=int, default=1, metavar='N',
                     help='model updates per simulator step (default: 1)')
@@ -158,11 +158,26 @@ try:
 except Exception as e:
     pass
 
-with open(train_csv, 'w') as f:
-    f.write("episode,reward\n")
+try:
+    if not os.path.exists(train_csv):
+        with open(train_csv, 'w') as f:
+            f.write("episode,reward\n")
+    else:
+        with open(train_csv, 'a') as f:
+            f.write("-1,-1\n")
+except Exception as e:
+    pass
 
-with open(test_csv, 'w') as f:
-    f.write("episode,reward\n")
+try:
+    if not os.path.exists(test_csv):
+        with open(test_csv, 'w') as f:
+            f.write("episode,reward\n")
+    else:
+        with open(test_csv, 'a') as f:
+            f.write("-1,-1\n")
+except Exception as e:
+    pass
+
 
 #Tensorboard
 writer = SummaryWriter(experiment)
@@ -320,6 +335,9 @@ for i_episode in itertools.count(1):
         
 
         writer.add_scalar('avg_reward/test', avg_reward, i_episode)
+
+        with open(test_csv, 'a') as f:
+            f.write(f"{i_episode},{avg_reward}\n")
 
         print("----------------------------------------")
         # print("Config: {}|{}|{} Test Episodes: {}, Avg. Reward: {}".format(args.config,args.cup_flag,args.adaptive,episodes, round(avg_reward, 2)))
